Cardio/requirements.txt
astroid==2.2.5
certifi==2020.6.20
colorama==0.4.1
cycler==0.10.0
isort==4.3.20
kiwisolver==1.2.0
lazy-object-proxy==1.4.1
matplotlib==3.3.1
mccabe==0.6.1
numpy==1.19.1
Pillow==7.2.0
pylint==2.3.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scipy==1.5.2
six==1.15.0
typed-ast==1.3.5
wincertstore==0.2
yapf==0.28.0
